chore: update public.packages according to ABBS before OOT retirement
The public.packages table used to contain the list of packages in the
ABBS tree, however this table is no longer used due to package site
moved to a separate database.

We now perform a scan of the ABBS tree, and update the package list to
table public.packages, to allow the out-of-tree query to work.